pyproject.toml
[project]
name = "xbox-soarm-teleop"
version = "0.1.0"
description = "Xbox controller teleoperation for SO-ARM100/101 using LeRobot IK"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Rick Anderson", email = "rick.rickanderson@gmail.com"},
]
dependencies = [
    "lerobot[kinematics]",
    "inputs>=0.5",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "mujoco>=3.4.0",
    "scservo-sdk",
    "scipy>=1.15.3",
    "pygame>=2.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff>=0.4.0",
    "pre-commit",
]
joycon = [
    "evdev>=1.6",
]
pygame = [
    "pygame>=2.5.0",
]
rerun = [
    "rerun-sdk>=0.21",
]

[project.scripts]
ik-smoke = "xbox_soarm_teleop.cli.ik_smoke:main"
mujoco-ik-check = "xbox_soarm_teleop.cli.mujoco_ik_check:main"
analyze-joint-diag = "xbox_soarm_teleop.cli.analyze_joint_diag:main"
diagnose-robot = "xbox_soarm_teleop.cli.diagnose_robot:main"
joint-rom-test = "xbox_soarm_teleop.cli.joint_rom_test:main"
xbox-joint-diagnostic = "xbox_soarm_teleop.cli.xbox_joint_diagnostic:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xbox_soarm_teleop"]

[tool.ruff]
line-length = 104
target-version = "py310"
exclude = [".git", ".venv", "__pycache__", "dist", "build"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["xbox_soarm_teleop"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.uv.workspace]
members = [
    "third_party/ftservo_python",
]

[tool.uv.sources]
scservo-sdk = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
